Create data_extractors and move Clip - WIP
